datastore_core_server (and related children...) updated to use new classes datastore_auth and datastore_database

+++ datastore/__init__.py
@@ -11,11 +11,6 @@ import xmlrpclib
 import inspect
 import importlib
 
-# Lets go multi-thread (Threaded mix-in) 
-class AsyncXMLRPCServer(SocketServer.ThreadingMixIn,
-                        SimpleXMLRPCServer.SimpleXMLRPCServer): pass
-
-
 # data acces mode (r/w)
 AUTHMODE_NONE = 0
 AUTHMODE_READ = 1
@@ -31,11 +26,9 @@ VARTYPE_ANY = "A"
 # file
 VARTYPE_FILE = "F"
 
-#if len(sys.argv) > 1:
-#    port = int(sys.argv[1])
-
 # Set up logging
 #logging.basicConfig(level=logging.DEBUG)
+
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
 handler = logging.handlers.SysLogHandler(address = '/dev/log')
@@ -43,8 +36,187 @@ formatter = logging.Formatter('%(module)s.%(funcName)s: %(message)s')
 handler.setFormatter(formatter)
 log.addHandler(handler)
 
+class datastore_auth(object):
+	def __init__(self, ldapserver, userdn, groupdn, binddn="", bindpass=""):
+		self.ldapserver = ldapserver
+		self.userdn = userdn
+		self.groupdn = groupdn
+		self.binddn = binddn
+		self.bindpass = bindpass
+		self._initialized = False
+
+	def initialize(self):
+		if self._initialized:
+			return True
+
+		try:
+			self.ld = ldap.initialize(self.ldapserver)
+			self.ld.protocol_version = ldap.VERSION3
+			self._initialized = True
+		except ldap.LDAPError, e:
+			self._initialized = False
+
+		return self._initialized
+
+	def test_credentials(self, username, userpass):
+		if not self.initialize():
+			return False
+
+		dn = "uid="+username+","+self.userdn
+		try:
+			self.ld.simple_bind_s(dn, userpass)
+			return True
+		except ldap.LDAPError, e:
+			return False
+		
+	def list_groups(self, username):
+		glist = []
+		if self.initialize():
+			searchScope = ldap.SCOPE_SUBTREE
+			retrieveAttributes = [ 'cn' ] 
+			searchFilter = "memberuid="+username
+			if self.binddn and self.bindpass:
+				self.ld.simple_bind_s(self.binddn, self.bindpass)
+
+                        ldap_result_id = self.ld.search(self.groupdn, searchScope, searchFilter, retrieveAttributes)
+                        while 1:
+       	                        result_type, result_data = self.ld.result(ldap_result_id, 0)
+               	                if (result_data == []):
+                               	        break
+                       	        else:
+					if result_type == ldap.RES_SEARCH_ENTRY:
+						glist.append(result_data[0][1]['cn'][0])
+
+		return glist
+
+
+# Create class datastore_database
+class datastore_database(object):
+	def __init__(self,host,user,passwd,dbname):
+		self.dbhost=host 
+		self.dbuser=user
+		self.dbpass=passwd
+		self.dbname=dbname
+		self._initialized_conn = False
+		self._initialized_cursor = False 		
+
+	def init_db(self):
+		#connect
+		if not self._initialized_conn:
+			try:
+				self.db_conn = MySQLdb.connect(self.dbhost, self.dbuser, self.dbpass, self.dbname)
+				self._initialized_conn = True
+			except:
+				self._initialized_conn = False
+		return self._initialized_conn
+
+	
+	def close(self, db_conn):
+		if self._initilized:
+			self.db_conn.close()			
+	
+	def cursor_execute(self, query):
+		if not self._initialized_cursor:
+			if self.init_db():
+				try:
+					self.cursor= self.db_conn.cursor()
+					self._initialized_cursor = True
+
+				except:
+					self._initialized_cursor = False 
+		if not self._initialized_cursor:
+			return False
+
+		try:	
+			self.cursor.execute(query)
+			return True
+		except:
+			return False
+
+
+	def delete(self, namespace, varname):
+		if self.cursor_execute("DELETE FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname)):
+			self.db_conn.commit()
+			return True
+
+		self.db_conn.rollback()
+		return False
+
+	def update(self, namespace, varname, varvalue, vartype=VARTYPE_STRING):
+		succeed = False
+		if ( vartype == VARTYPE_PASSWORD ):
+			varvalue = base64.b64encode(varvalue)
+
+		if not self.cursor_execute("SELECT * FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname)):
+			return False
+		if len(self.cursor.fetchall()) > 0 :
+			# do update
+			succeed = self.cursor_execute("UPDATE varvalues SET varvalue='%s', vartype='%s' WHERE namespace='%s' AND varname='%s';" % (varvalue, vartype, namespace, varname))
+		else:
+			# do insert
+			succeed = self.cursor_execute("INSERT INTO varvalues(namespace, varname, varvalue, vartype) VALUES ('%s', '%s', '%s', '%s');" % (namespace, varname, varvalue, vartype))
+			
+		if succeed:
+			self.db_conn.commit()
+			return True
+		else:
+			# Rollback in case there is any error
+			self.db_conn.rollback()
+			return False
+
+
+	def read(self, namespace, varname):
+		if self.cursor_execute("SELECT varvalue, vartype FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname)):
+			# get_result = (cur.fetchall())[0][0]
+			# comprobar el numero de filas leidas por si no hay
+			try:
+				first_row =(self.cursor.fetchall())[0]
+				# prepare data
+				if ( first_row[1] == VARTYPE_PASSWORD ):
+					get_data = base64.b64decode(first_row[0])
+				else:
+					get_data = first_row[0]
+			except:
+				get_data = ""
+		return get_data
+
+
+	def get_filepath(self, namespace):
+		if self.cursor_execute("SELECT filepath FROM filepaths WHERE namespace='%s';" % (namespace)):
+			filepath = (self.cursor.fetchall())[0][0]
+		else:
+			filepath = ""
+		return filepath
+	
+	def test_auth(self, group_list, namespace, varname, accesslevel, filemode=False):
+		auth_user = False
+		if filemode:
+			selectindex = 1
+		else:
+			selectindex = 0
+
+		#search in database
+		for g in group_list:
+			# do SQL query
+			if self.cursor.execute("SELECT authvar,authfile FROM auth WHERE username='%s' AND namespace='%s' AND (varname='%s' OR varname='' OR varname IS NULL);" % (g, namespace, varname)):
+				# print all the first cell of all the rows
+				for row in self.cursor.fetchall():
+					if ( row[selectindex] >= accesslevel ):
+						auth_user = True
+		return auth_user
+
+
+# Lets go multi-thread (Threaded mix-in) 
+class AsyncXMLRPCServer(SocketServer.ThreadingMixIn,
+                        SimpleXMLRPCServer.SimpleXMLRPCServer): pass
+
 # Create class Server with exposed methods
 class datastore_core_server(object):
+	def __init__(self, ds_auth, ds_database, debug_mode = False):
+		self.ds_auth = ds_auth
+		self.ds_database = ds_database	
+		self.debug_mode = debug_mode
+
 	def _get_groups(self, username, userpass):
 		valid_user = False
 		log_message = 'user validation error'
@@ -55,14 +227,10 @@ class datastore_core_server(object):
 			return glist
 		else:
 			# validate credentials in ldap ...
-			try:
-				ld=ldap.initialize(self.ldapserver)
-				ld.protocol_version = ldap.VERSION3	
-				dn = "uid="+username+","+self.userdn
-				ld.simple_bind_s(dn,userpass)
+			if self.ds_auth.test_credentials(username, userpass):
 				valid_user = True
 				glist = [ 'anonymous', username ]
-			except ldap.LDAPError, e:
+			else:
 				valid_user = False	
 		if valid_user:
 			log_message = 'user validation successfully'
@@ -72,27 +240,12 @@ class datastore_core_server(object):
 		if self.debug_mode:
 			log.debug('validate_user(%s): '+log_message, username)
 		# get group list
-		searchScope = ldap.SCOPE_SUBTREE
-		retrieveAttributes = [ 'cn' ] 
-		searchFilter = "memberuid="+username
-		try:
-			if self.binddn and self.bindpass:
-				ld.simple_bind_s(self.binddn, self.bindpass)
-
-			ldap_result_id = ld.search(self.groupdn, searchScope, searchFilter, retrieveAttributes)
-			while 1:
-				result_type, result_data = ld.result(ldap_result_id, 0)
-				if (result_data == []):
-					break
-				else:
-					if result_type == ldap.RES_SEARCH_ENTRY:
-						#glist.append(result_data)
-						glist.append('@'+result_data[0][1]['cn'][0])
-			log_message = 'search successfully'
-
-		except ldap.LDAPError, e:
-			log_message = 'search error'
-
+		log_message = 'LDAP groups found:'
+		ldap_groups = self.ds_auth.list_groups(username)
+		for g in ldap_groups:
+			log_message = log_message + ' ' + g
+			glist.append('@'+g)
+			
 		if self.debug_mode:
 			log.debug('get_groups(%s): '+log_message, username)
 		return glist
@@ -104,33 +257,14 @@ class datastore_core_server(object):
 		return self._test_auth(username, userpass, namespace, varname, accesslevel, True)
 
 	def _test_auth(self, username, userpass, namespace, varname, accesslevel, filemode=False):
-		auth_user = False
-		if filemode:
-			selectindex = 1
+		auth_user = self.ds_database.test_auth(self._get_groups(username,userpass), namespace, varname, accesslevel, filemode)
+		if auth_user:
+			log_message = 'Access granted for user (%s)'
 		else:
-			selectindex = 0
-
-		#search in database
-		try:
-			# connect
-			db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)	
-			cur = db.cursor() 
-			for g in self._get_groups(username, userpass):
-				# do SQL query
-				cur.execute("SELECT authvar,authfile FROM auth WHERE username='%s' AND namespace='%s' AND (varname='%s' OR varname='' OR varname IS NULL);" % (g, namespace, varname))
-				# print all the first cell of all the rows
-				for row in cur.fetchall() :
-					if ( row[selectindex] >= accesslevel ):
-						auth_user = True
-						break
-			cur.close()
-			db.close()
-			log_message = "Successfully access to database"
-		except:
-			log_message = "Error accessing database"
+			log_message = "Access denied for user (%s)"
 				
 		if self.debug_mode:
-			log.debug('database access: '+log_message)
+			log.debug('database access: '+log_message,username)
 
 		return auth_user
 
@@ -153,33 +287,10 @@ class datastore_core_server(object):
 		if not self._is_valid_data(varvalue, vartype):
 			log_message = "put_value: invalid data"
 		elif self._test_auth_var(username, userpass, namespace, varname, AUTHMODE_WRITE):
-			# prepare data
-			if ( vartype == VARTYPE_PASSWORD ):
-				varvalue = base64.b64encode(varvalue)
-
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor()
-				try:
-					cur.execute("SELECT * FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname))
-					if len(cur.fetchall()) > 0 :
-						# do update
-						cur.execute("UPDATE varvalues SET varvalue='%s', vartype='%s' WHERE namespace='%s' AND varname='%s';" % (varvalue, vartype, namespace, varname))
-					else:
-						# do insert
-						cur.execute("INSERT INTO varvalues(namespace, varname, varvalue, vartype) VALUES ('%s', '%s', '%s', '%s');" % (namespace, varname, varvalue, vartype))
-			
-					db.commit()
-				except:
-					# Rollback in case there is any error
-					db.rollback()
-
-				cur.close()
-				db.close()
+			if self.ds_database.update(self, namespace, varname, varvalue, vartype=VARTYPE_STRING):
 				log_message = "successfully put_value"
 				put_result = True
-			except:
+			else:
 				log_message = "error in  put_value"
 
 		if self.debug_mode:
@@ -191,21 +302,14 @@ class datastore_core_server(object):
 		del_result = False
 		log_message = "del_value: not authorized"
 		if self._test_auth_var(username, userpass, namespace, varname, AUTHMODE_WRITE):
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor()
-				cur.execute("DELETE FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname))
+			if self.ds_database.delete(namespace, varname):
 				log_message = "del_value: success"
-				cur.close()
-				db.close()
 				del_result = True
-			except:
+			else:
 				log_message = "del_value: error deleting value"
 				
 		if self.debug_mode:
 			log.debug(log_message)
-				
 			
 		return del_result
 
@@ -213,23 +317,9 @@ class datastore_core_server(object):
 		log_message = "get_value: not authorized"
 		get_data = ""
 		if self._test_auth_var(username, userpass, namespace, varname, AUTHMODE_READ):
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor() 
-				cur.execute("SELECT varvalue, vartype FROM varvalues WHERE namespace='%s' AND varname='%s';" % (namespace, varname))
-				# get_result = (cur.fetchall())[0][0]
-				first_row =(cur.fetchall())[0]
-				# prepare data
-				if ( first_row[1] == VARTYPE_PASSWORD ):
-					get_data = base64.b64decode(first_row[0])
-				else:
-					get_data = first_row[0]
-
-				cur.close()
-				db.close()
+			if self.ds_database.read(namespace, varname):
 				log_message = "successfully get_value"
-			except:
+			else:
 				log_message = "error in  get_value"
 
 		if self.debug_mode:
@@ -242,14 +332,8 @@ class datastore_core_server(object):
 		log_message = "put_file: not authorized"
 		if self._test_auth_file(username, userpass, namespace, fname, AUTHMODE_WRITE):
 			# get path to store files
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor()
-				cur.execute("SELECT filepath FROM filepaths WHERE namespace='%s';" % (namespace))
-				filepath = (cur.fetchall())[0][0]
-				cur.close()
-				db.close()
+			filepath = self.ds_database.get_filepath(namespace)
+			if filepath:
 				# do update
 				try:
 					with open(filepath+"/"+os.path.basename(fname), "wb") as handle:
@@ -261,7 +345,7 @@ class datastore_core_server(object):
 				except:
 					log_message = "error writing file"
 
-			except:
+			else:
 				log_message = "error reading path"
 
 		if self.debug_mode:
@@ -274,14 +358,8 @@ class datastore_core_server(object):
 		log_message = "del_file: not authorized"
 		if self._test_auth_file(username, userpass, namespace, fname, AUTHMODE_WRITE):
 			# get path to store files
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor()
-				cur.execute("SELECT filepath FROM filepaths WHERE namespace='%s';" % (namespace))
-				filepath = (cur.fetchall())[0][0]
-				cur.close()
-				db.close()
+			filepath = self.ds_database.get_filepath(namespace)
+			if filepath:
 				# do update
 				try:
 					os.remove(filepath+"/"+os.path.basename(fname))
@@ -291,7 +369,7 @@ class datastore_core_server(object):
 				except:
 					log_message = "error deleting file"
 
-			except:
+			else:
 				log_message = "error reading path"
 
 		if self.debug_mode:
@@ -304,14 +382,8 @@ class datastore_core_server(object):
 		binary_data = xmlrpclib.Binary('')
 		if self._test_auth_file(username, userpass, namespace, fname, AUTHMODE_WRITE):
 			# get path where files are stored
-			try:
-				# connect
-				db = MySQLdb.connect(host=dbhost, user=dbuser, passwd=dbpass, db=dbname)
-				cur = db.cursor()
-				cur.execute("SELECT filepath FROM filepaths WHERE namespace='%s';" % (namespace))
-				filepath = (cur.fetchall())[0][0]
-				cur.close()
-				db.close()
+			filepath = self.ds_database.get_filepath(namespace)
+			if filepath:
 				# do update
 				try:
 
@@ -323,7 +395,7 @@ class datastore_core_server(object):
 				except:
 					log_message = "error reading file"
 
-			except:
+			else:
 				log_message = "error reading path"
 
 		if self.debug_mode:
@@ -360,7 +432,8 @@ class datastore_plugin(datastore_core_server):
 		return super(datastore_plugin,self).get_file(username, userpass, self.__class__.name__, fname)
 
 class datastore_server(datastore_core_server):
-	def __init__(self,plugin_path=""):
+	def __init__(self, ds_auth, ds_database, plugin_path="", debug_mode = False):
+		super(datastore_core_server,self).__init__(ds_auth, ds_database, debug_mode)
 		if plugin_path:
 			if plugin_path not in sys.path:
 				sys.path.append(plugin_path)
@@ -369,6 +442,6 @@ class datastore_server(datastore_core_server):
 					cname = os.path.splitext(file)[0]
 					cmodule = importlib.import_module(cname)
 					constructor = getattr(cmodule, cname)
-					setattr( self, cname,constructor() )
+					setattr( self, cname,constructor(ds_auth, ds_database, debug_mode) )
